requirements.txt
flask
pymongo
requests
pinecone-client
mongoengine
PyJWT
python-dotenv
flask-cors
pytesseract
pdf2image
docx2txt
whisper
youtube_transcript_api
vimeo_dl
langchain 
openai 
tiktoken
pinecone
langchain-community



